+++ scripts/test/tearDown.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-# Copyright 2022 Crown Copyright
+# Copyright 2022-2023 Crown Copyright
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -16,21 +16,19 @@
 set -e
 # Tears down a Sleeper instance and associated System Testing framework
 
-this_dir=$(cd $(dirname $0) && pwd)
-PROJECT_ROOT=$(dirname $(dirname ${this_dir}))
-pushd ${PROJECT_ROOT}/java
-VERSION=$(mvn -q -DforceStdout help:evaluate -Dexpression=project.version)
-DEPLOY_SCRIPTS_DIR="${PROJECT_ROOT}/scripts/deploy"
-GENERATED_DIR="${PROJECT_ROOT}/scripts/generated"
+SCRIPTS_DIR=$(cd "$(dirname "$0")" && cd .. && pwd)
+VERSION=$(cat "$SCRIPTS_DIR/templates/version.txt")
+DEPLOY_SCRIPTS_DIR="${SCRIPTS_DIR}/deploy"
+GENERATED_DIR="${SCRIPTS_DIR}/generated"
 INSTANCE_PROPERTIES=${GENERATED_DIR}/instance.properties
-INSTANCE_ID=$(fgrep sleeper.id ${INSTANCE_PROPERTIES} | cut -d'=' -f2)
-CONFIG_BUCKET=$(cat ${GENERATED_DIR}/configBucket.txt)
-TABLE_BUCKET=$(cat ${GENERATED_DIR}/tableBucket.txt)
-QUERY_BUCKET=$(cat ${GENERATED_DIR}/queryResultsBucket.txt)
-RETAIN_INFRA=`grep sleeper.retain.infra.after.destroy ${INSTANCE_PROPERTIES} | cut -d'=' -f2 | awk '{print tolower($0)}'`
+INSTANCE_ID=$(grep -F sleeper.id "${INSTANCE_PROPERTIES}" | cut -d'=' -f2)
+CONFIG_BUCKET=$(cat "${GENERATED_DIR}/configBucket.txt")
+TABLE_BUCKET=$(cat "${GENERATED_DIR}/tableBucket.txt")
+QUERY_BUCKET=$(cat "${GENERATED_DIR}/queryResultsBucket.txt")
+RETAIN_INFRA=$(grep sleeper.retain.infra.after.destroy "${INSTANCE_PROPERTIES}" | cut -d'=' -f2 | awk '{print tolower($0)}')
 
 echo "RETAIN_INFRA:${RETAIN_INFRA}"
-echo "PROJECT_ROOT:${PROJECT_ROOT}"
+echo "SCRIPTS_DIR:${SCRIPTS_DIR}"
 echo "GENERATED_DIR:${GENERATED_DIR}"
 echo "INSTANCE_PROPERTIES:${INSTANCE_PROPERTIES}"
 echo "CONFIG_BUCKET:${CONFIG_BUCKET}"
@@ -38,28 +36,46 @@ echo "TABLE_BUCKET:${TABLE_BUCKET}"
 echo "QUERY_BUCKET:${QUERY_BUCKET}"
 echo "DEPLOY_SCRIPTS_DIR:${DEPLOY_SCRIPTS_DIR}"
 
-pushd ${PROJECT_ROOT}
+source "${SCRIPTS_DIR}/functions/timeUtils.sh"
+START_TIME=$(record_time)
+echo "Started at $(recorded_time_str "$START_TIME")"
 
 echo "Pausing the system"
-java -cp java/clients/target/clients-*-utility.jar "sleeper.status.update.PauseSystem" ${INSTANCE_ID}
+java -cp "${SCRIPTS_DIR}/jars/clients-${VERSION}-utility.jar" "sleeper.status.update.PauseSystem" "${INSTANCE_ID}"
+
+END_PAUSE_TIME=$(record_time)
+echo "Pause finished at $(recorded_time_str "$END_PAUSE_TIME"), took $(elapsed_time_str "$START_TIME" "$END_PAUSE_TIME")"
 
 if [[ "${RETAIN_INFRA}" == "false" ]]; then
   echo "Removing all data from config, table and query results buckets"
   # Don't fail script if buckets don't exist
-  aws s3 rm s3://${CONFIG_BUCKET} --recursive || true
-  aws s3 rm s3://${TABLE_BUCKET} --recursive || true
-  aws s3 rm s3://${QUERY_BUCKET} --recursive || true
+  aws s3 rm "s3://${CONFIG_BUCKET}" --recursive || true
+  aws s3 rm "s3://${TABLE_BUCKET}" --recursive || true
+  aws s3 rm "s3://${QUERY_BUCKET}" --recursive || true
 fi
 
+END_CLEAR_BUCKETS_TIME=$(record_time)
+echo "Clear buckets finished at $(recorded_time_str "$END_CLEAR_BUCKETS_TIME"), took $(elapsed_time_str "$END_PAUSE_TIME" "$END_CLEAR_BUCKETS_TIME")"
+
 echo "Running cdk destroy to remove the system"
-cdk -a "java -cp java/system-test/target/system-test-*-utility.jar sleeper.systemtest.cdk.SystemTestApp" \
-destroy -c testpropertiesfile=${INSTANCE_PROPERTIES} "*"
+cdk -a "java -cp ${SCRIPTS_DIR}/jars/system-test-${VERSION}-utility.jar sleeper.systemtest.cdk.SystemTestApp" \
+destroy --force -c testpropertiesfile="${INSTANCE_PROPERTIES}" "*"
+
+END_CDK_DESTROY_TIME=$(record_time)
+echo "CDK destroy finished at $(recorded_time_str "$END_CDK_DESTROY_TIME"), took $(elapsed_time_str "$END_CLEAR_BUCKETS_TIME" "$END_CDK_DESTROY_TIME")"
 
 echo "Removing the Jars bucket and docker containers"
-${DEPLOY_SCRIPTS_DIR}/removeUploads.sh ${INSTANCE_PROPERTIES}
-popd
+"${DEPLOY_SCRIPTS_DIR}/removeUploads.sh" "${INSTANCE_PROPERTIES}"
 
 echo "Removing generated files"
-rm -r ${GENERATED_DIR}
+rm -r "${GENERATED_DIR:?}"/*
 
 echo "Successfully torn down"
+
+FINISH_TIME=$(record_time)
+echo "Started at $(recorded_time_str "$START_TIME")"
+echo "Pause finished at $(recorded_time_str "$END_PAUSE_TIME"), took $(elapsed_time_str "$START_TIME" "$END_PAUSE_TIME")"
+echo "Clear buckets finished at $(recorded_time_str "$END_CLEAR_BUCKETS_TIME"), took $(elapsed_time_str "$END_PAUSE_TIME" "$END_CLEAR_BUCKETS_TIME")"
+echo "CDK destroy finished at $(recorded_time_str "$END_CDK_DESTROY_TIME"), took $(elapsed_time_str "$END_CLEAR_BUCKETS_TIME" "$END_CDK_DESTROY_TIME")"
+echo "Removing buckets & files finished at $(recorded_time_str "$FINISH_TIME"), took $(elapsed_time_str "$END_CDK_DESTROY_TIME" "$FINISH_TIME")"
+echo "Overall, took $(elapsed_time_str "$START_TIME" "$FINISH_TIME")"
